Add MTU test code
We wrote this up back when we were having an issue with openser on EC2
because we neglected to notice that the MTU was being set to 9000 by
default.  The effect of that was that UDP packets were being
fragmented making the UDP checksums incorrect.  The packets were
therefore being dropped by the receiver.